formatflowed didn't work, so parsing broke there. fixing that.
formatflowed expects bytes, but we are passing it a string and quietly
catching the exception (iow we're not converting it).

While not a good thing, it doesn't impact message handling terribly much
except for in the permalink generation for generators that make use of
the parsed body. Therefore, I am proposing we adopt the conversion fix
in Foal and leave as is in the current pony mail.